+++ Src/FluentAssertions/Equivalency/Execution/CollectionMemberOptionsDecorator.cs
@@ -79,8 +79,6 @@ internal class CollectionMemberOptionsDecorator : IEquivalencyOptions
 
     public bool IgnoreCase => inner.IgnoreCase;
 
-    public bool IgnoreAllNewlines => inner.IgnoreAllNewlines;
-
     public bool IgnoreNewlineStyle => inner.IgnoreNewlineStyle;
 
     public ITraceWriter TraceWriter => inner.TraceWriter;