+++ Src/FluentAssertions/Equivalency/SelfReferenceEquivalencyOptions.cs
@@ -91,7 +91,6 @@ public abstract class SelfReferenceEquivalencyOptions<TSelf> : IEquivalencyOptio
         IgnoreLeadingWhitespace = defaults.IgnoreLeadingWhitespace;
         IgnoreTrailingWhitespace = defaults.IgnoreTrailingWhitespace;
         IgnoreCase = defaults.IgnoreCase;
-        IgnoreAllNewlines = defaults.IgnoreAllNewlines;
         IgnoreNewlineStyle = defaults.IgnoreNewlineStyle;
 
         ConversionSelector = defaults.ConversionSelector.Clone();
@@ -192,8 +191,6 @@ public abstract class SelfReferenceEquivalencyOptions<TSelf> : IEquivalencyOptio
     public bool IgnoreTrailingWhitespace { get; private set; }
 
     public bool IgnoreCase { get; private set; }
-
-    public bool IgnoreAllNewlines { get; private set; }
 
     public bool IgnoreNewlineStyle { get; private set; }
 
@@ -735,18 +732,6 @@ public abstract class SelfReferenceEquivalencyOptions<TSelf> : IEquivalencyOptio
     public TSelf IgnoringCase()
     {
         IgnoreCase = true;
-        return (TSelf)this;
-    }
-
-    /// <summary>
-    /// Instructs the comparison to ignore all newlines when comparing <see langword="string" />s.
-    /// </summary>
-    /// <remarks>
-    /// Enabling this option will remove all newlines from the strings before comparing them.
-    /// </remarks>
-    public TSelf IgnoringAllNewlines()
-    {
-        IgnoreAllNewlines = true;
         return (TSelf)this;
     }