+++ Src/FluentAssertions/Primitives/StringAssertions.cs
@@ -471,7 +471,6 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
             new StringWildcardMatchingStrategy
             {
                 IgnoreCase = options.IgnoreCase,
-                IgnoreAllNewlines = options.IgnoreAllNewlines,
                 IgnoreNewlineStyle = options.IgnoreNewlineStyle,
             },
             because, becauseArgs);
@@ -597,7 +596,6 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
             new StringWildcardMatchingStrategy
             {
                 IgnoreCase = options.IgnoreCase,
-                IgnoreAllNewlines = options.IgnoreAllNewlines,
                 IgnoreNewlineStyle = options.IgnoreNewlineStyle,
                 Negate = true
             },
@@ -1998,7 +1996,6 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     /// <remarks>
     /// When <see cref="IEquivalencyOptions.IgnoreLeadingWhitespace"/> is set, whitespace is removed from the start of the <paramref name="value"/>.<br />
     /// When <see cref="IEquivalencyOptions.IgnoreTrailingWhitespace"/> is set, whitespace is removed from the end of the <paramref name="value"/>.<br />
-    /// When <see cref="IEquivalencyOptions.IgnoreAllNewlines"/> is set, all newlines are removed from the <paramref name="value"/>.<br />
     /// When <see cref="IEquivalencyOptions.IgnoreNewlineStyle"/> is set, all newlines (<c>\r\n</c> and <c>\r</c>) are replaced with <c>\n</c> in the <paramref name="value"/>.<br />
     /// </remarks>
     private static string ApplyStringSettings(string value, IEquivalencyOptions options)
@@ -2011,11 +2008,6 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         if (options.IgnoreTrailingWhitespace)
         {
             value = value.TrimEnd();
-        }
-
-        if (options.IgnoreAllNewlines)
-        {
-            value = value.RemoveNewLines();
         }
 
         if (options.IgnoreNewlineStyle)