+++ Tests/FluentAssertions.Specs/Collections/CollectionAssertionSpecs.BeEquivalentTo.cs
@@ -193,17 +193,6 @@ public partial class CollectionAssertionSpecs
         }
 
         [Fact]
-        public void Can_ignore_all_newlines_while_comparing_collections_of_strings()
-        {
-            // Arrange
-            var actual = new[] { "first", "\rA\nB\r\nC\n", "last" };
-            var expectation = new[] { "first", "ABC", "last" };
-
-            // Act / Assert
-            actual.Should().BeEquivalentTo(expectation, o => o.IgnoringAllNewlines());
-        }
-
-        [Fact]
         public void Can_ignore_newline_style_while_comparing_collections_of_strings()
         {
             // Arrange