+++ Tests/FluentAssertions.Specs/Primitives/ObjectAssertionSpecs.BeEquivalentTo.cs
@@ -40,17 +40,6 @@ public partial class ObjectAssertionSpecs
         }
 
         [Fact]
-        public void Can_ignore_all_newlines_while_comparing_objects_with_string_properties()
-        {
-            // Arrange
-            var actual = new { foo = "\rA\nB\r\nC\n" };
-            var expectation = new { foo = "ABC" };
-
-            // Act / Assert
-            actual.Should().BeEquivalentTo(expectation, o => o.IgnoringAllNewlines());
-        }
-
-        [Fact]
         public void Can_ignore_newline_style_while_comparing_objects_with_string_properties()
         {
             // Arrange