+++ Tests/FluentAssertions.Specs/Primitives/StringAssertionSpecs.BeEquivalentTo.cs
@@ -73,17 +73,6 @@ public partial class StringAssertionSpecs
         }
 
         [Fact]
-        public void Can_ignore_all_newlines_while_comparing_strings_to_be_equivalent()
-        {
-            // Arrange
-            string actual = "\rA\nB\r\nC\n";
-            string expect = "\r\nAB\nC";
-
-            // Act / Assert
-            actual.Should().BeEquivalentTo(expect, o => o.IgnoringAllNewlines());
-        }
-
-        [Fact]
         public void Can_ignore_newline_style_while_comparing_strings_to_be_equivalent()
         {
             // Arrange
@@ -254,20 +243,6 @@ public partial class StringAssertionSpecs
 
             // Act
             Action act = () => actual.Should().NotBeEquivalentTo(expect, o => o.IgnoringTrailingWhitespace());
-
-            // Assert
-            act.Should().Throw<XunitException>();
-        }
-
-        [Fact]
-        public void Can_ignore_all_newlines_while_comparing_strings_to_not_be_equivalent()
-        {
-            // Arrange
-            string actual = "\rA\nB\r\nC\n";
-            string expect = "\nA\r\nB\n\nC";
-
-            // Act
-            Action act = () => actual.Should().NotBeEquivalentTo(expect, o => o.IgnoringAllNewlines());
 
             // Assert
             act.Should().Throw<XunitException>();